NHIS 2019 Survey EDA
This looks at the CDC National Health Interview Survey data from the sample adult population for the year 2019. The data variable definitions can be found in the Codebook for 2019, but I've included relevant ones in the notebook. This exploration focuses on demographics, treatments, comorbidities and other insights that may be beneficial to our group.